Whisp/backend/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
python-dotenv==1.0.0
flask-jwt-extended==4.5.3
psycopg2-binary==2.9.9
sib_api_v3_sdk==7.6.0